requirements.txt
streamlit
python-dotenv
requests
langchain
faiss-cpu
openai
google-api-python-client
google-auth
google-auth-oauthlib
langchain-community
